feat: support aarch64-linux builds
IDA 9.3 ships a native Linux ARM64 build but the SDK does not include
lib/arm64_linux_gcc_64 stub libraries. Link directly against the runtime
.so files in the install dir on this target, and define __ARM__ for the
SDK headers (pro.h gates NULL_VA_LIST and va_list handling on it).

On aarch64, C char is unsigned; bindgen emits u8 for op_t::offb/offo/
specflag* and inf_get_nametype/strlit_zeroes where x86_64 saw i8. Change
the wrapper return types to u8 with a portable cast — these are byte
offsets, flag bits, and counts, never negative.

Adds requires_local_ida_install() so consumers can fail fast with a
useful message instead of a cryptic linker error when the install is
missing on this target.

+++ idalib/src/insn.rs
@@ -153,12 +153,17 @@ impl Operand {
         OperandFlags::from_bits_retain(self.inner.flags)
     }
 
-    pub fn offb(&self) -> i8 {
-        self.inner.offb
+    // SDK declares offb/offo/specflag* as `char`, which is signed on x86_64
+    // but unsigned on aarch64. Cast to u8 for a stable, semantically correct API.
+    // The cast is a no-op on aarch64, hence the allow.
+    #[allow(clippy::unnecessary_cast)]
+    pub fn offb(&self) -> u8 {
+        self.inner.offb as u8
     }
 
-    pub fn offo(&self) -> i8 {
-        self.inner.offo
+    #[allow(clippy::unnecessary_cast)]
+    pub fn offo(&self) -> u8 {
+        self.inner.offo as u8
     }
 
     pub fn n(&self) -> usize {
@@ -256,33 +261,37 @@ impl Operand {
         }
     }
 
-    pub fn processor_specific_flag1(&self) -> Option<i8> {
+    #[allow(clippy::unnecessary_cast)]
+    pub fn processor_specific_flag1(&self) -> Option<u8> {
         if self.is_processor_specific() {
-            Some(self.inner.specflag1)
+            Some(self.inner.specflag1 as u8)
         } else {
             None
         }
     }
 
-    pub fn processor_specific_flag2(&self) -> Option<i8> {
+    #[allow(clippy::unnecessary_cast)]
+    pub fn processor_specific_flag2(&self) -> Option<u8> {
         if self.is_processor_specific() {
-            Some(self.inner.specflag2)
+            Some(self.inner.specflag2 as u8)
         } else {
             None
         }
     }
 
-    pub fn processor_specific_flag3(&self) -> Option<i8> {
+    #[allow(clippy::unnecessary_cast)]
+    pub fn processor_specific_flag3(&self) -> Option<u8> {
         if self.is_processor_specific() {
-            Some(self.inner.specflag3)
+            Some(self.inner.specflag3 as u8)
         } else {
             None
         }
     }
 
-    pub fn processor_specific_flag4(&self) -> Option<i8> {
+    #[allow(clippy::unnecessary_cast)]
+    pub fn processor_specific_flag4(&self) -> Option<u8> {
         if self.is_processor_specific() {
-            Some(self.inner.specflag4)
+            Some(self.inner.specflag4 as u8)
         } else {
             None
         }
